# Improve performance by doing formatting ourselves