--- a/stack.yaml
+++ b/stack.yaml
@@ -1,0 +1,2 @@
+resolver: lts-8.9
+packages: ['.']