This PR replaces use of the `formatting` library with custom functions. In my testing, this reduces CUID generation time by about half (see #3). Before:

```
time                 3.758 μs   (3.699 μs .. 3.827 μs)
                     0.998 R²   (0.997 R² .. 0.999 R²)
mean                 3.781 μs   (3.728 μs .. 3.835 μs)
```

After:

```
time                 1.919 μs   (1.892 μs .. 1.948 μs)
                     0.998 R²   (0.997 R² .. 0.999 R²)
mean                 1.916 μs   (1.878 μs .. 1.954 μs)
```

I also added a `stack.yaml` to version control so that developers using Stack wouldn’t have to run `stack init` every time.