@crabmusket 2017-04-17
Hey this is awesome! Thanks for this. You've reminded me I need to fix up my Travis configuration :p. I'll take a look over this and get back to you within a day or two :)