@bdesham 2017-04-17
Hey, happy to do it!

If you’re willing to switch the Travis builds from Cabal to Stack, you could copy my setup from [this repository](https://github.com/bdesham/pinboard-notes-backup), which in turn is copied from [this repository](https://github.com/tfausak/octane). (Building only with Stack is fine for my application, but I don’t know if it’s what you’d want for a library.)