@crabmusket 2017-04-18
I decided to stay with `cabal`-based testing for now, though in the future testing with Stack might be a nice idea. Stuff's been fixed in `master` FYI. I'm pretty happy with how your code looks, and thanks for documenting everything well! I'm happy to merge as soon as we sort out the `.gitignore` :p