@bdesham 2017-04-18
I rebased my work on top of the latest `master`; all of the tests pass.

What’s wrong with the `.gitignore`? 😄