@crabmusket 2017-04-19
It looks like it contains an absolute path. See my comment. I'm just going to merge this and then fix it :). Your contribution is greatly appreciated!