@crabmusket 2017-04-28
@bdesham FYI I just published version 1.2.0.1 to hackage